fix: adding unitProcessing to Installation

+++ PRIO/src/Modules/Measuring/Measurements/Infra/EF/Models/Volume.cs
@@ -6,8 +6,8 @@ namespace PRIO.src.Modules.Measuring.Equipments.Infra.EF.Models
     public class Volume : BaseModel
     {
         public DateTime? DHA_MEDICAO_039 { get; set; }
-        public double? DHA_MED_DECLARADO_039 { get; set; }
-        public double? DHA_MED_REGISTRADO_039 { get; set; }
+        public decimal? DHA_MED_DECLARADO_039 { get; set; }
+        public decimal? DHA_MED_REGISTRADO_039 { get; set; }
         [JsonIgnore]
         public Measurement? Measurement { get; set; }
     }
